fixed: changed int64 on Ids an Ref to int

+++ app/api/v1/schemas/course.go
@@ -1,42 +1,42 @@
 package schemas
 
 type CourseRes struct {
-	Id          int64  `json:"id"`
-	Version     int64  `json:"version"`
+	Id          int    `json:"id"`
+	Version     int    `json:"version"`
 	Description string `json:"description"`
 	Number      string `json:"number"`
 }
 
 type CourseResStudent struct {
-	Id          int64               `json:"id"`
-	Version     int64               `json:"version"`
+	Id          int                 `json:"id"`
+	Version     int                 `json:"version"`
 	Description string              `json:"description"`
 	Number      string              `json:"number"`
 	Exams       []CourseExamStudent `json:"exams"`
 }
 
 type CourseResTeacher struct {
-	Id          int64               `json:"id"`
-	Version     int64               `json:"version"`
+	Id          int                 `json:"id"`
+	Version     int                 `json:"version"`
 	Description string              `json:"description"`
 	Number      string              `json:"number"`
 	Exams       []CourseExamTeacher `json:"exams"`
 }
 
 type CourseReqPut struct {
-	Version     int64  `json:"version"`
+	Version     int    `json:"version"`
 	Description string `json:"description"`
 	Number      string `json:"number"`
-	ModuleRef   int64  `json:"integer"`
+	ModuleRef   int    `json:"integer"`
 }
 
 type CourseReqPost struct {
 	Description string `json:"description" validate:"required"`
 	Number      string `json:"number"`
-	ModuleRef   int64  `json:"moduleRef"`
+	ModuleRef   int    `json:"moduleRef"`
 }
 
 type CourseFilter struct {
 	Modules  []Module `json:"modules"`
-	Versions []int64  `json:"versions"`
+	Versions []int    `json:"versions"`
 }
